+++ LLIO2025/src/Models/index.ts
@@ -129,4 +129,5 @@ export interface OutlookEventRaw {
 export interface OutlookEvent extends Omit<OutlookEventRaw, 'start' | 'end'> {
   start: Date;
   end: Date;
+  submitted: boolean;
 }
